requirements/base.txt
fastapi[all]
sqlmodel
geopy
googlemaps
PyJWT
magic-admin